--- src/SFA.DAS.EAS.Portal/Application/Adapters/IAdapter.cs
@@ -1,8 +1,0 @@
-﻿
-namespace SFA.DAS.EAS.Portal.Application.Adapters
-{
-    public interface IAdapter<T1, T2>
-    {
-        T2 Convert(T1 input);
-    }
-}
