--- src/SFA.DAS.EAS.Portal/Application/Commands/ICommand.cs
@@ -1,6 +1,0 @@
-﻿namespace SFA.DAS.EAS.Portal.Application.Commands
-{
-    public interface ICommand
-    {
-    }
-}
